LayoutTests/platform/wpe/fast/mediastream/captureStream/canvas3d-expected.txt

CONSOLE MESSAGE: InvalidStateError: Video element has no video frame
CONSOLE MESSAGE: InvalidStateError: Video element has no video frame



Harness Error (FAIL), message = InvalidStateError: Video element has no video frame

FAIL captureStream with 3d context drawing - not buffered initially promise_test: Unhandled rejection with value: "no rvfc for 1"
FAIL captureStream with 3d context drawing - buffered initially promise_test: Unhandled rejection with value: "no rvfc for 1"
